requirements.txt
python-dotenv
droidrun[google,dev]
llama-index-llms-google-genai
pydantic
jinja2
nest_asyncio